requirements.txt
django
djangorestframework
django-rest-knox
django-health-check
faker
faker_vehicle